Move InteractsWithKeyboard methods to InteractsWithPlaywright
